added seed and replaced job model
