+++ server/package.json
@@ -5,7 +5,8 @@
     "main": "server.js",
     "scripts": {
         "start": "node server.js",
-        "watch": "nodemon server.js"
+        "watch": "nodemon server.js",
+        "seed": "node seeds/seed.js"
     },
     "keywords": [],
     "author": "",
